--- a/brave/src/BraveApp.swift
+++ b/brave/src/BraveApp.swift
@@ -184,7 +184,9 @@ class BraveApp {
     }
 
     static func showErrorAlert(title title: String,  error: String) {
-        UIAlertView(title: title, message: error, delegate: nil, cancelButtonTitle: "Close").show()
+        delay(0) {
+            UIAlertView(title: title, message: error, delegate: nil, cancelButtonTitle: "Close").show()
+        }
     }
 
     static func statusBarHeight() -> CGFloat {