--- a/brave/src/webfilters/TrackingProtection.swift
+++ b/brave/src/webfilters/TrackingProtection.swift
@@ -73,6 +73,10 @@ class TrackingProtection {
 }
 
 extension TrackingProtection: NetworkDataFileLoaderDelegate {
+    func fileLoader(loader: NetworkDataFileLoader, convertDataBeforeWriting data: NSData, etag: String?) {
+        loader.finishWritingToDisk(data, etag: etag)
+    }
+
     func fileLoader(_: NetworkDataFileLoader, setDataFile data: NSData?) {
         parser.setDataFile(data)
     }