deserialization times on 5c:
JSON: ~7s
NSKeyedArchiver: ~5s
FastCoder: ~1.5s

serialization is ~10s with all of them